+++ src/Chunkyard/Options/CopyOptions.cs
@@ -13,10 +13,10 @@ namespace Chunkyard.Options
             DestinationRepository = destinationRepository;
         }
 
-        [Option('s', "source-repository", Required = true, HelpText = "The source repository")]
+        [Option('s', "source", Required = true, HelpText = "The source repository path")]
         public string SourceRepository { get; }
 
-        [Option('d', "destination-repository", Required = true, HelpText = "The destination repository")]
+        [Option('d', "destination", Required = true, HelpText = "The destination repository path")]
         public string DestinationRepository { get; }
     }
 }