+++ src/Chunkyard/Options/RestoreOptions.cs
@@ -19,19 +19,19 @@ namespace Chunkyard.Options
             Overwrite = overwrite;
         }
 
-        [Option('r', "repository", Required = true, HelpText = "The repository")]
+        [Option('r', "repository", Required = true, HelpText = "The repository path")]
         public string Repository { get; }
 
         [Option('d', "directory", Required = false, HelpText = "The directory to restore into", Default = ".")]
         public string Directory { get; }
 
-        [Option('i', "include", Required = false, HelpText = "The include fuzzy pattern")]
+        [Option('i', "include", Required = false, HelpText = "The fuzzy pattern for files to include")]
         public string IncludeFuzzy { get; }
 
-        [Option('l', "log-position", Required = false, HelpText = "The log position", Default = Cli.LatestLogPosition)]
+        [Option('s', "snapshot", Required = false, HelpText = "The snapshot ID", Default = Cli.LatestLogPosition)]
         public int LogPosition { get; }
 
-        [Option('o', "overwrite", Required = false, HelpText = "If files should be overwritten", Default = false)]
+        [Option("overwrite", Required = false, HelpText = "If files should be overwritten", Default = false)]
         public bool Overwrite { get; }
     }
 }