+++ src/Chunkyard/Options/ShowOptions.cs
@@ -15,13 +15,13 @@ namespace Chunkyard.Options
             IncludeFuzzy = includeFuzzy;
         }
 
-        [Option('r', "repository", Required = true, HelpText = "The repository")]
+        [Option('r', "repository", Required = true, HelpText = "The repository path")]
         public string Repository { get; }
 
-        [Option('l', "log-position", Required = false, HelpText = "The log position", Default = Cli.LatestLogPosition)]
+        [Option('s', "snapshot", Required = false, HelpText = "The snapshot ID", Default = Cli.LatestLogPosition)]
         public int LogPosition { get; }
 
-        [Option('i', "include", Required = false, HelpText = "The include fuzzy pattern")]
+        [Option('i', "include", Required = false, HelpText = "The fuzzy pattern for files to include")]
         public string IncludeFuzzy { get; }
     }
 }